+++ bec_lib/bec_lib/bl_state_manager.py
@@ -6,6 +6,7 @@ from collections.abc import Mapping, Sequence
 from inspect import Parameter, Signature
 from typing import TYPE_CHECKING, Any, TypedDict
 
+import yaml
 from pydantic import BaseModel
 from rich.console import Console
 from rich.table import Table
@@ -14,10 +15,13 @@ from rich.tree import Tree
 
 from bec_lib import bl_states, messages
 from bec_lib.endpoints import MessageEndpoints
+from bec_lib.logger import bec_logger
 
 if TYPE_CHECKING:
     from bec_lib.bl_states import BeamlineStateConfig
     from bec_lib.client import BECClient
+
+logger = bec_logger.logger
 
 
 def build_signature_from_model(model: BaseModel, skip: set[str] | None = None) -> Signature:
@@ -203,6 +207,89 @@ class BeamlineStateManager:
         self._connector.register(
             MessageEndpoints.available_beamline_states(), cb=self._on_state_update
         )
+
+    def load_from_config(
+        self,
+        config_path: str | None = None,
+        config_dict: dict | None = None,
+        flush: bool = True,
+        skip_existing: bool = False,
+    ) -> None:
+        """
+        Load a state configuration from a YAML file or a dictionary. If None or both are provided,
+        an error will be raised. Configs must adhere to the following structure:
+
+        ``` yaml
+        <state_name>:
+          bl_state_class: <class_name>
+          config:
+            <config_key>: <config_value>
+        ...
+        ```
+
+        Args:
+            config_path (str | None): The path to the YAML configuration file.
+            config_dict (dict | None): A dictionary containing the configuration.
+            flush (bool): If True, existing states in the manager will be cleared before loading new ones.
+            skip_existing (bool): If True, existing states in the manager will be skipped during loading.
+        """
+        self._check_inputs(config_path=config_path, config_dict=config_dict)
+        config_dict = self._load_config(config_path=config_path, config_dict=config_dict)
+        # Check first if the config is valid before clearing the existing states
+        configs = self._parse_config(config_dict=config_dict)
+        if flush:
+            self.clear_all()
+        for config in configs:
+            try:
+                self.add(config, skip_existing=skip_existing)
+            except TimeoutError:
+                logger.warning(f"Timeout while waiting for state {config.name} to be active.")
+
+    def _check_inputs(self, config_path: str | None, config_dict: dict | None) -> None:
+        """Utility method to check that either config_path or config_dict is provided, but not both."""
+        if (config_path is None and config_dict is None) or (
+            config_path is not None and config_dict is not None
+        ):
+            raise ValueError("Either config_path or config_dict must be provided, but not both.")
+
+    def _load_config(self, config_path: str | None, config_dict: dict | None) -> dict:
+        """Utility method to load the configuration from a YAML file or return the provided dictionary."""
+        if config_path:
+            with open(config_path, "r", encoding="utf-8") as f:
+                loaded_config = yaml.safe_load(f)
+            if loaded_config is None:
+                raise ValueError("Config file is empty.")
+            return loaded_config
+        if config_dict is None:
+            raise ValueError("config_dict must be provided when config_path is not set.")
+        return config_dict
+
+    def _parse_config(self, config_dict: dict) -> list[bl_states.BeamlineStateConfig]:
+        """Parse the configuration dictionary into a list of BeamlineStateConfig instances. Raise if invalid."""
+        parsed_configs: list[bl_states.BeamlineStateConfig] = []
+        for state_name, entry in config_dict.items():
+            state_config_class = self._resolve_config_class(entry["bl_state_class"])
+            config = entry["config"]
+            if not isinstance(config, dict):
+                raise ValueError(f"Config for state {state_name!r} must be a dictionary.")
+            if "name" in config and config["name"] != state_name:
+                raise ValueError(
+                    f"Config name {config['name']!r} does not match top-level state name {state_name!r}."
+                )
+            config = {**config, "name": state_name}
+            parsed_configs.append(state_config_class(**config))
+        return parsed_configs
+
+    def _resolve_config_class(
+        self, bl_state_class: str | type[bl_states.BeamlineStateConfig]
+    ) -> type[bl_states.BeamlineStateConfig]:
+        """Resolve the configuration class for a given beamline state class name or type."""
+        if isinstance(bl_state_class, str):
+            resolved_class = _state_class_for_state_type(bl_state_class)
+        else:
+            resolved_class = bl_state_class
+        config = resolved_class.CONFIG_CLASS
+        return config
 
     @property
     def ready(self) -> bool:
